// src/app/shared/model/employeedetails.model.ts
export interface EmployeeDetailsModel {
    empId: Number;
    empCode: String;
    empName: String;
    designation: String;
    fName: String;
    mName: String;
    dob: String;
    address: String;
    phoneNumber: String;
    mobileNumber: String;
    emailId: String;
    id: String;
    idNumber: String;
    joinDate: String;
    reportingOfficer: String;
    remarks: String;
    empStatus: String;
    resignDate: String;
    division: String;
    submitTime: Date;
    ipaddress: String;
    workorderTo: Date;
    agencyName:string;

    docFileNameImg: string;
    docTypeContentImg: string;

    docFileNameId: string;
    docTypeContentId: string;

    docFileNameAddressProof: string;
    docTypeContentAddressProof: string;
}